Tests: block state tests
